+++ src/GiapTech.LangCenter.LMS.Infrastructure/Persistence/AppDbContext.cs
@@ -20,6 +20,7 @@ public class AppDbContext(DbContextOptions<AppDbContext> options, ICurrentTenant
     public DbSet<HoSoGiaoVien> HoSoGiaoViens => Set<HoSoGiaoVien>();
     public DbSet<HoSoHocVien> HoSoHocViens => Set<HoSoHocVien>();
     public DbSet<HoSoNhanVien> HoSoNhanViens => Set<HoSoNhanVien>();
+    public DbSet<PhongBan> PhongBans => Set<PhongBan>();
 
     public DbSet<Quyen> Quyens => Set<Quyen>();
     public DbSet<QuyenChucNang> QuyenChucNangs => Set<QuyenChucNang>();